requirement.txt
flask
spotipy
python-dotenv
pandas
transformers
torch
numpy<2.0.0